[refrac] Rename Scripts dir
- Bump version to 0.2.0
