+++ setup.py
@@ -12,7 +12,7 @@ with codecs.open('requirements.txt', 'r', 'utf8') as reader:
 
 setup(
     name='MedicalQA',
-    version='0.01.0',
+    version='0.2.0',
     packages=find_packages(),
     url='https://github.com/Santosh-Gupta/MedicalQA',
     license='MIT',
@@ -21,8 +21,9 @@ setup(
     description='BERT in TF2.0 for Medical QA info retrieval + GPT2 for answer generation',
     long_description='None so far',
     long_description_content_taype='text/markdown',
-    python_requires = '>=3.6.0',
-    install_requires=['numpy', 'matplotlib', 'tensorflow==2.0.0-alpha0', 'tensorflow-gpu==2.0.0-alpha0', 'Keras', 'keras-pos-embd==0.9.0', 'keras-transformer==0.21.0', 'tqdm', 'faiss', 'sklearn', 'pycurl', 'six', 'argparse'], 
+    python_requires='>=3.5.0',
+    install_requires=['numpy', 'matplotlib', 'tensorflow==2.0.0-alpha0', 'tensorflow-gpu==2.0.0-alpha0', 'Keras',
+                      'keras-pos-embd==0.9.0', 'keras-transformer==0.21.0', 'tqdm', 'faiss', 'sklearn', 'pycurl', 'six', 'argparse'],
     classifiers=(
         "Programming Language :: Python :: 2.7",
         "Programming Language :: Python :: 3.6",
@@ -30,7 +31,3 @@ setup(
         "Operating System :: OS Independent",
     ),
 )
-
-
-
-
